requirements.txt
Django==1.8.12
mock==2.0.0
requests==2.9.1
responses==0.5.1
simplejson==3.8.2
versiontag==1.0.3
djangorestframework==3.3.3
python-dateutil==2.5.2
